* misc: removed some typed-punned warnings by replacing all IsMinListEmpty()   calls by IsListEmpty() equivalents. These warnings are only issues by GCC   4.1.0+ compilers which is now the new default compiler suite for the OS4   builds.

+++ src/YAM_rexx_rxif.h
@@ -220,14 +220,14 @@ struct rxd_folderinfo
                 char *folder;
         } arg;
         struct {
-                long *number;
+                int *number;
                 char *name;
                 char *path;
-                long *total;
-                long *new;
-                long *unread;
-                long *size;
-                long *type;
+                int *total;
+                int *new;
+                int *unread;
+                LONG *size;
+                int *type;
         } res;
 };
 
@@ -284,7 +284,7 @@ struct rxd_getselected
                 char *var, *stem;
         } arg;
         struct {
-                long **num;
+                int **num;
         } res;
 };
